+++ .NET/CoreFundamentals/CoreFundamentals.Data/SqlRestaurantData.cs
@@ -1,0 +1,59 @@
+﻿using CoreFundamentals.Core;
+using Microsoft.EntityFrameworkCore;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace CoreFundamentals.Data
+{
+    public class SqlRestaurantData : IRestaurantData
+    {
+        private readonly CoreFundamentalsDbContext db;
+
+        public SqlRestaurantData(CoreFundamentalsDbContext db)
+        {
+            this.db = db;
+        }
+
+        public int Commit()
+        {
+            return db.SaveChanges();
+        }
+
+        public Restaurant Create(Restaurant newRestaurant)
+        {
+            db.Add(newRestaurant);
+            return newRestaurant;
+        }
+
+        public Restaurant Delete(int id)
+        {
+            var restaurant = GetById(id);
+            if(restaurant != null)
+            {
+                db.Restaurants.Remove(restaurant);
+            }
+            return restaurant;
+        }
+
+        public Restaurant GetById(int id)
+        {
+            return db.Restaurants.Find(id);
+        }
+
+        public IEnumerable<Restaurant> GetRestaurantsByName(string name)
+        {
+            var query = from r in db.Restaurants
+                        where r.Name.StartsWith(name) || string.IsNullOrEmpty(name)
+                        orderby r.Name
+                        select r;
+            return query;
+        }
+
+        public Restaurant Update(Restaurant updatedRestaurant)
+        {
+            var entity = db.Restaurants.Attach(updatedRestaurant);
+            entity.State = EntityState.Modified;
+            return updatedRestaurant;
+        }
+    }
+}
