+++ .NET/CoreFundamentals/CoreFundamentals/Startup.cs
@@ -6,6 +6,7 @@ using CoreFundamentals.Data;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.HttpsPolicy;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -24,8 +25,16 @@ namespace CoreFundamentals
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddDbContextPool<CoreFundamentalsDbContext>(options =>
+            {
+                options.UseSqlServer(Configuration.GetConnectionString("CoreFundamentalsDb"));
+            });
+
             //Only for development env, in a real env this would create data corruption  
-            services.AddSingleton<IRestaurantData, InMemoryRestaurantData>();
+            //services.AddSingleton<IRestaurantData, InMemoryRestaurantData>();
+            
+            //With EF
+            services.AddScoped<IRestaurantData, SqlRestaurantData>();
 
             services.AddRazorPages();
         }
