+++ .NET/CoreFundamentals/CoreFundamentals/appsettings.json
@@ -7,5 +7,8 @@
     }
   },
   "AllowedHosts": "*",
-  "Message":  "Hello from appsettings!"
+  "Message": "Hello from appsettings!",
+  "ConnectionStrings": {
+    "CoreFundamentalsDb": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=CoreFundamentals;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
+  }
 }
